requirements.txt

bottle
bson
pymongo
nltk
pandas
numpy
python-dotenv
cython
scipy

